move over test file and populate unfinished tests
